requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
boto3==1.42.69
botocore==1.42.69
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dataclasses-json==0.6.7
filelock==3.25.2
Flask==3.1.0
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.12
langchain-aws==1.4.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.18
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.3
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==4.6.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.48
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
